requirements.txt
mysqlclient
pymysql
pandas
sqlalchemy
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.1.2
Django==1.11.6
django-appconf==1.0.2
django-axes==3.0.3
django-easysettings==1.1
django-password-policies-iplweb==0.4.4b1
django-pyodbc-azure==1.11.0.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
enum34==1.1.6
et-xmlfile==1.0.1
html5lib==0.9999999
httpie==0.9.9
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Keras==2.1.1
Markdown==2.6.9
MarkupSafe==1.0
networkx==2.0
olefile==0.44
protobuf==3.5.0.post1
Pygments==2.2.0
pyodbc==4.0.21
pyparsing==2.2.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
xlrd==1.1.0
wfastcgi
django-session-security
